test(gateway): lock Lazer subscription write fields (ENG-400)
Assert the Lazer websocket subscription frame requests price, confidence, exponent, emaPrice, emaConfidence, and feedUpdateTimestamp with formats:solana, deliveryFormat:json, and jsonBinaryEncoding:base64 — the properties required by the Pyth Pro adapter stateful write path.

Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
